- now set haveData in try{} instead of finally{} - added some logging (temporary) - replaced Vue.nextTick() with this.$nextTick() - removed some unneeded "ignore changes" - moved some "step validation" from created() to mounted() - app version = 3.4.2
