+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "business-create-ui",
-  "version": "3.4.1",
+  "version": "3.4.2",
   "private": true,
   "appName": "Create UI",
   "sbcName": "SBC Common Components",
