+++ src/views/Dissolution/DissolutionAffidavit.vue
@@ -5,8 +5,8 @@
 </template>
 
 <script lang="ts">
-import { Component, Mixins, Vue, Watch } from 'vue-property-decorator'
-import Affidavit from '@/components/Dissolution/Affidavit.vue'
+import { Component, Mixins, Watch } from 'vue-property-decorator'
+import Affidavit from '@/components/Dissolution/CompleteAffidavit.vue'
 import { RouteNames } from '@/enums'
 import { CommonMixin } from '@/mixins'
 import { Getter } from 'vuex-class'
@@ -26,7 +26,7 @@ export default class DissolutionAffidavit extends Mixins(CommonMixin) {
   private async scrollToInvalidComponent (): Promise<void> {
     if (this.getShowErrors && this.$route.name === RouteNames.DISSOLUTION_AFFIDAVIT) {
       // Scroll to invalid components.
-      await Vue.nextTick()
+      await this.$nextTick()
       const vid = this.getAffidavitStep.validationDetail.validationItemDetails
       const validFlags = this.buildValidFlags(vid)
       const componentIds = this.buildElementIds(vid)
@@ -38,7 +38,3 @@ export default class DissolutionAffidavit extends Mixins(CommonMixin) {
   }
 }
 </script>
-
-<style lang="scss" scoped>
-@import '@/assets/styles/theme.scss';
-</style>
