+++ src/views/Incorporation/IncorporationRules.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-import { Component, Mixins, Vue, Watch } from 'vue-property-decorator'
+import { Component, Mixins, Watch } from 'vue-property-decorator'
 import { Getter } from 'vuex-class'
 import { CreateRulesIF } from '@/interfaces'
 import { CommonMixin } from '@/mixins'
@@ -26,7 +26,7 @@ export default class IncorporationRules extends Mixins(CommonMixin) {
   private async scrollToInvalidComponent (): Promise<void> {
     if (this.getShowErrors && this.$route.name === RouteNames.INCORPORATION_RULES) {
       // scroll to invalid components
-      await Vue.nextTick()
+      await this.$nextTick()
       const vid = this.getCreateRulesStep.validationDetail.validationItemDetails
       const validFlags = this.buildValidFlags(vid)
       const componentIds = this.buildElementIds(vid)
@@ -38,7 +38,3 @@ export default class IncorporationRules extends Mixins(CommonMixin) {
   }
 }
 </script>
-
-<style lang="scss" scoped>
-@import '@/assets/styles/theme.scss';
-</style>
