+++ tests/unit/UploadAffidavit.spec.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import { shallowWrapperFactory, wrapperFactory } from '../jest-wrapper-factory'
-import Affidavit from '@/components/Dissolution/Affidavit.vue'
+import Affidavit from '@/components/Dissolution/CompleteAffidavit.vue'
 import { DissolutionResources } from '@/resources'
 import { CorpTypeCd } from '@bcrs-shared-components/corp-type-module'
